Ejercicio 4 y edicion de detalles del 5

+++ backend/controllers/studentsController.php
@@ -10,6 +10,8 @@
 */
 
 require_once("./repositories/students.php");
+require_once("./repositories/studentsSubjects.php");
+
 
 // Para GET (usamos la variable superglobal $_GET):
 //https://www.php.net/manual/es/language.variables.superglobals.php
@@ -86,16 +88,22 @@ function handlePut($conn)
 function handleDelete($conn) 
 {
     $input = json_decode(file_get_contents("php://input"), true);
-
-    $result = deleteStudent($conn, $input['id']);
-    if ($result['deleted'] > 0) 
-    {
-        echo json_encode(["message" => "Eliminado correctamente"]);
-    } 
-    else 
-    {
+   if ( !ExisteAlumno($conn,$input['id'])) {
+       $result = deleteStudent($conn, $input['id']);
+       if ($result['deleted'] > 0) 
+       {
+         echo json_encode(["message" => "Eliminado correctamente"]);
+       } 
+      else 
+       {
         http_response_code(500);
         echo json_encode(["error" => "No se pudo eliminar"]);
+       }
     }
+    else {
+        http_response_code(200);
+        echo json_encode(["message" => "No se puede eliminar al estudiante por tener asignaciones"]);
+    }
+        
 }
 ?>
